+++ src/java/src/iaik/pkcs/pkcs11/objects/PublicKey.java
@@ -164,8 +164,7 @@ public class PublicKey extends Key {
         newObject = KEAPublicKey.getInstance(session, objectHandle);
       } else if (keyType.equals(Key.KeyType.X9_42_DH)) {
         newObject = X942DHPublicKey.getInstance(session, objectHandle);
-      } else if ((keyType.longValue()
-              & KeyType.VENDOR_DEFINED.longValue()) != 0L) {
+      } else if ((keyType & KeyType.VENDOR_DEFINED) != 0L) {
         newObject = getUnknownPublicKey(session, objectHandle);
       } else {
         newObject = getUnknownPublicKey(session, objectHandle);
